+++ src/frontend/js/hooks/useCourses/index.spec.tsx
@@ -13,7 +13,8 @@ import JoanieSessionProvider from 'contexts/SessionContext/JoanieSessionProvider
 import { createTestQueryClient } from 'utils/test/createTestQueryClient';
 import { User } from 'types/User';
 import { Deferred } from 'utils/test/deferred';
-import { useCourses, TeacherCourseSearchFilters, CourseStatusFilter, CourseTypeFilter } from '.';
+import { PaginatedResourceQuery } from 'types/Joanie';
+import { useCourses } from '.';
 
 jest.mock('utils/context', () => ({
   __esModule: true,
@@ -25,7 +26,7 @@ jest.mock('utils/context', () => ({
 
 interface RenderUseCoursesProps extends PropsWithChildren {
   user?: User;
-  filters?: TeacherCourseSearchFilters;
+  filters?: PaginatedResourceQuery;
 }
 const renderUseCourses = ({ user, filters }: RenderUseCoursesProps) => {
   const Wrapper = ({ children }: PropsWithChildren) => (
@@ -52,10 +53,7 @@ describe('hooks/useCourses', () => {
 
   it('fetch all courses', async () => {
     const responseDeferred = new Deferred();
-    fetchMock.get(
-      'https://joanie.endpoint/api/v1.0/courses/?status=all&type=all',
-      responseDeferred.promise,
-    );
+    fetchMock.get('https://joanie.endpoint/api/v1.0/courses/', responseDeferred.promise);
 
     const user = UserFactory().one();
     const { result } = renderUseCourses({ user });
@@ -85,68 +83,5 @@ describe('hooks/useCourses', () => {
     expect(result.current.states.updating).toBeUndefined();
     expect(result.current.states.isLoading).toBe(false);
     expect(result.current.states.error).toBeUndefined();
-  });
-
-  it('fetch with filter "incoming"', async () => {
-    const courseRuns = CourseListItemFactory().many(3);
-    fetchMock.get('https://joanie.endpoint/api/v1.0/courses/?status=incoming&type=all', courseRuns);
-
-    const user = UserFactory().one();
-    const filters: TeacherCourseSearchFilters = {
-      status: CourseStatusFilter.INCOMING,
-      type: CourseTypeFilter.ALL,
-    };
-
-    const { result } = renderUseCourses({ user, filters });
-    await waitFor(() => {
-      expect(result.current.states.fetching).toBe(false);
-    });
-    const calledUrls = fetchMock.calls().map((call) => call[0]);
-    expect(calledUrls).toContain(
-      'https://joanie.endpoint/api/v1.0/courses/?status=incoming&type=all',
-    );
-    expect(JSON.stringify(result.current.items)).toBe(JSON.stringify(courseRuns));
-  });
-
-  it('fetch with filter "ongoing"', async () => {
-    const courseRuns = CourseListItemFactory().many(3);
-    fetchMock.get('https://joanie.endpoint/api/v1.0/courses/?status=ongoing&type=all', courseRuns);
-
-    const user = UserFactory().one();
-    const filters: TeacherCourseSearchFilters = {
-      status: CourseStatusFilter.ONGOING,
-      type: CourseTypeFilter.ALL,
-    };
-
-    const { result } = renderUseCourses({ user, filters });
-    await waitFor(() => {
-      expect(result.current.states.fetching).toBe(false);
-    });
-    const calledUrls = fetchMock.calls().map((call) => call[0]);
-    expect(calledUrls).toContain(
-      'https://joanie.endpoint/api/v1.0/courses/?status=ongoing&type=all',
-    );
-    expect(JSON.stringify(result.current.items)).toBe(JSON.stringify(courseRuns));
-  });
-
-  it('fetch with filter "archived"', async () => {
-    const courseRuns = CourseListItemFactory().many(3);
-    fetchMock.get('https://joanie.endpoint/api/v1.0/courses/?status=archived&type=all', courseRuns);
-
-    const user = UserFactory().one();
-    const filters: TeacherCourseSearchFilters = {
-      status: CourseStatusFilter.ARCHIVED,
-      type: CourseTypeFilter.ALL,
-    };
-
-    const { result } = renderUseCourses({ user, filters });
-    await waitFor(() => {
-      expect(result.current.states.fetching).toBe(false);
-    });
-    const calledUrls = fetchMock.calls().map((call) => call[0]);
-    expect(calledUrls).toContain(
-      'https://joanie.endpoint/api/v1.0/courses/?status=archived&type=all',
-    );
-    expect(JSON.stringify(result.current.items)).toBe(JSON.stringify(courseRuns));
   });
 });