+++ src/frontend/js/utils/test/mockPaginatedResponse.ts
@@ -1,0 +1,14 @@
+import { PaginatedResponse } from 'types/Joanie';
+
+export const mockPaginatedResponse: <Data>(
+  results?: Data[],
+  totalCount?: number,
+  haveNextPage?: boolean,
+) => PaginatedResponse<Data> = (results = [], totalCount = 0, haveNextPage = true) => {
+  return {
+    count: totalCount,
+    next: haveNextPage ? 'next' : null,
+    prev: null,
+    results,
+  };
+};